Fiexed bug for deploy 2

+++ router.mjs
@@ -26,8 +26,14 @@ router.get('/', async (req, res) => {
     if (req.query.user) {
       const userfilter = {username: req.query.user}
       const user = await User.findOne(userfilter).exec();
-      filter.user = user._id;
+      if (user !== null) {
+        filter.user = user._id;
+      }
+      else{
+        filter.user = null;
+      }
     }
+
     // const posts = await Post.find(filter).sort('-createdAt').populate('user').exec();
     const posts = await Post.find(filter)
       .sort('-createdAt')
@@ -158,7 +164,6 @@ router.get('/post/:slug', async (req, res) => {
   else{
     const postSlug = req.params.slug;
     const TargetPost = await Post.findOne({slug: postSlug}).populate('user').exec();
-    console.log(TargetPost);
     if (TargetPost.user._id == req.session.user._id) {
       res.render('my-post-slug', {post: TargetPost});
     }
@@ -168,7 +173,11 @@ router.get('/post/:slug', async (req, res) => {
 
 // comment route
 router.post('/post/:slug/comment', async (req, res) => {
-    const postSlug = req.params.slug;
+    if (!req.session.user) {
+      res.redirect('/login');
+    }
+    else{
+      const postSlug = req.params.slug;
     const body = sanitize(req.body.body);
     if (!body) {
       res.redirect('/post/' + postSlug);
@@ -185,20 +194,31 @@ router.post('/post/:slug/comment', async (req, res) => {
       res.redirect('/post/' + postSlug);
       });
     }
+  }
 });
 
 // delete post route
 router.get('/post/:slug/delete', async (req, res) => {
-    const postSlug = req.params.slug;
-    await Post.findOneAndDelete({slug: postSlug}).exec();
-    res.redirect('/');
+    if (!req.session.user) {
+      res.redirect('/login');
+    }
+    else{
+      const postSlug = req.params.slug;
+      await Post.findOneAndDelete({slug: postSlug}).exec();
+      res.redirect('/');
+    }
 });
 
 //edit post route
 router.get('/post/:slug/edit', async (req, res) => {
+  if (!req.session.user) {
+    res.redirect('/login');
+  }
+  else{
     const postSlug = req.params.slug;
     const TargetPost = await Post.findOne({slug: postSlug}).exec();
     res.render('edit', {title: TargetPost.title, body: TargetPost.body, topic: TargetPost.topic});
+  }
 });
 
 router.post('/post/:slug/edit', async (req, res) => {
